django-ecommerce/requirements.txt
asgiref==3.5.2
aws-secretsmanager-caching==1.1.1.5
botocore==1.29.7
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
Markdown==3.4.1
Pillow==9.3.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stripe==4.2.0
tzdata==2022.6
urllib3==1.26.12
